+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## 3.20.0 - 2020-07-05
+
 ### Added
 
 - SASL config fields added to `amqp_1` input and output.